pyproject.toml
[tool.poetry]
name = "ssl-backend"
version = "0.1.0"
description = ""
authors = ["tsdocode <tsdocode@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
alembic = "^1.11.1"
uvicorn = "^0.22.0"
starlette = "0.27.0"
fastapi = "^0.97.0"
python-nginx = "^1.5.7"
psycopg2 = "^2.9.6"
python-dotenv = "^1.0.0"
python-on-whales = "^0.61.0"

[tool.poetry.scripts]
ssl-backend = "src.main:run"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
